requirements.txt
fastapi==0.74.1
uvicorn==0.17.5
black==22.1.0
python-decouple==3.6
asyncpg==0.25.0
SQLAlchemy==1.4.31
